TW-12435 - logger: Always-on OTLP stopwatch with 200ms threshold
